+++ myems-admin/app/controllers/settings/space/spacestore.controller.js
@@ -13,7 +13,8 @@ app.controller('SpaceStoreController', function(
     $scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
 
     $scope.getAllSpaces = function() {
-    SpaceService.getAllSpaces(function (response) {
+    let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+    SpaceService.getAllSpaces(headers, function (response) {
       if (angular.isDefined(response.status) && response.status === 200) {
         $scope.spaces = response.data;
       } else {
@@ -48,7 +49,8 @@ app.controller('SpaceStoreController', function(
 
 	$scope.getStoresBySpaceID = function(id) {
     $scope.spacestores=[];
-    SpaceStoreService.getStoresBySpaceID(id, function (response) {
+    let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+    SpaceStoreService.getStoresBySpaceID(id, headers, function (response) {
       				if (angular.isDefined(response.status) && response.status === 200) {
       					$scope.spacestores = $scope.spacestores.concat(response.data);
       				} else {
@@ -124,7 +126,8 @@ app.controller('SpaceStoreController', function(
 	$scope.getAllStores();
 
   $scope.refreshSpaceTree = function() {
-    SpaceService.getAllSpaces(function (response) {
+    let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+    SpaceService.getAllSpaces(headers, function (response) {
       if (angular.isDefined(response.status) && response.status === 200) {
         $scope.spaces = response.data;
       } else {