+++ myems-admin/app/controllers/settings/space/spaceworkingcalendar.controller.js
@@ -14,7 +14,8 @@ app.controller('SpaceWorkingCalendarController', function(
     $scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
 
     $scope.getAllSpaces = function() {
-    SpaceService.getAllSpaces(function (response) {
+    let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+    SpaceService.getAllSpaces(headers, function (response) {
       if (angular.isDefined(response.status) && response.status === 200) {
         $scope.spaces = response.data;
       } else {
@@ -49,8 +50,9 @@ app.controller('SpaceWorkingCalendarController', function(
     };
 
 	$scope.getWorkingCalendarsBySpaceID = function(id) {
-		$scope.spaceworkingcalendars=[];
-      SpaceWorkingCalendarService.getWorkingCalendarsBySpaceID(id, function (response) {
+	  $scope.spaceworkingcalendars=[];
+      let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+      SpaceWorkingCalendarService.getWorkingCalendarsBySpaceID(id, headers, function (response) {
             if (angular.isDefined(response.status) && response.status === 200) {
               $scope.spaceworkingcalendars = response.data;
             }
@@ -123,7 +125,8 @@ app.controller('SpaceWorkingCalendarController', function(
     $scope.getAllWorkingCalendars();
 
     $scope.refreshSpaceTree = function() {
-    SpaceService.getAllSpaces(function (response) {
+    let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+    SpaceService.getAllSpaces(headers, function (response) {
       if (angular.isDefined(response.status) && response.status === 200) {
         $scope.spaces = response.data;
       } else {