+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added access control to equipment in myems-api, myems-admin and myems-web
 - added access control to combined equipment in myems-api, myems-admin and myems-web
 - added access control to meter, offline meter and virtual meter in myems-api, myems-admin and myems-web
+- added access control to space in myems-api, myems-admin and myems-web
+
 ### Changed
 - 
 ### Fixed